Remove the need for an overlay
nvidia_x11 was overrode in an overlay. It is no longer the case and the
override is now explicit.

This will allows the creation of a `nixGL.nix` file callable with
`callPackage`. This change is a preliminary work for inclusion in
`nixpkgs`, as requested in #16.

It will also highly improve the use on nixGL in other project because
user won't have to pass a non initialized `nixpkgs`.