fix: find nvidia icd files thank to wildcard
Close #96.

It now uses a wildcard to find nvidia icd instead of a stack path. The
build does not fail anymore, but I'm not sure it works as expected. I
just tried to build the package and run the wrapper and dump the
generated path, and it generates the correct path.